backend/requirements.txt
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
firebase-admin==4.4.0
Flask==1.1.2
gcloud==0.17.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
httplib2==0.18.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
oauth2client==3.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dotenv==0.14.0
python-jwt==2.0.1
pytz==2020.1
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
